Made API better at defending against configurations for missing integrations
