+++ src/SpawnThat/Integrations/CLLC/Models/CllcBossAffix.cs
@@ -1,6 +1,4 @@
-﻿using CreatureLevelControl;
-
-namespace SpawnThat.Integrations.CLLC.Models;
+﻿namespace SpawnThat.Integrations.CLLC.Models;
 
 public enum CllcBossAffix
 {
@@ -12,22 +10,4 @@ public enum CllcBossAffix
     Elementalist,
     Enraged,
     Twin
-}
-
-internal static class CllcBossAffixExtensions
-{
-    public static BossAffix Convert(this CllcBossAffix affix)
-    {
-        return affix switch
-        {
-            CllcBossAffix.Reflective => BossAffix.Reflective,
-            CllcBossAffix.Shielded => BossAffix.Shielded,
-            CllcBossAffix.Mending => BossAffix.Mending,
-            CllcBossAffix.Summoner => BossAffix.Summoner,
-            CllcBossAffix.Elementalist => BossAffix.Elementalist,
-            CllcBossAffix.Enraged => BossAffix.Enraged,
-            CllcBossAffix.Twin => BossAffix.Twin,
-            _ => BossAffix.None,
-        };
-    }
 }
