+++ src/SpawnThat/Integrations/EpicLoot/Extensions/ILocalSpawnBuilderELConditionsExtensions.cs
@@ -11,7 +11,7 @@ public static class ILocalSpawnBuilderELConditionsExtensions
     {
         if (IntegrationManager.InstalledEpicLoot)
         {
-            builder.SetCondition(new ConditionNearbyPlayerCarryItemWithRarity(distanceToSearch, rarities));
+            SetConditionItemRarity(builder, distanceToSearch, rarities);
         }
 
         return builder;
@@ -21,7 +21,7 @@ public static class ILocalSpawnBuilderELConditionsExtensions
     {
         if (IntegrationManager.InstalledEpicLoot)
         {
-            builder.SetCondition(new ConditionNearbyPlayerCarryItemWithRarity(distanceToSearch, rarities));
+            SetConditionItemRarity(builder, distanceToSearch, rarities);
         }
 
         return builder;
@@ -31,7 +31,7 @@ public static class ILocalSpawnBuilderELConditionsExtensions
     {
         if (IntegrationManager.InstalledEpicLoot)
         {
-            builder.SetCondition(new ConditionNearbyPlayerCarryItemWithRarity(distanceToSearch, rarities));
+            SetConditionItemRarity(builder, distanceToSearch, rarities);
         }
 
         return builder;
@@ -41,17 +41,23 @@ public static class ILocalSpawnBuilderELConditionsExtensions
     {
         if (IntegrationManager.InstalledEpicLoot)
         {
-            builder.SetCondition(new ConditionNearbyPlayerCarryItemWithRarity(distanceToSearch, rarities));
+            SetConditionItemRarity(builder, distanceToSearch, rarities);
         }
 
         return builder;
     }
 
+    private static void SetConditionItemRarity(ILocalSpawnBuilder builder, int distance, IEnumerable<string> rarities) => builder.SetCondition(new ConditionNearbyPlayerCarryItemWithRarity(distance, rarities));
+    private static void SetConditionItemRarity(ILocalSpawnBuilder builder, int distance, params string[] rarities) => builder.SetCondition(new ConditionNearbyPlayerCarryItemWithRarity(distance, rarities));
+    private static void SetConditionItemRarity(ILocalSpawnBuilder builder, int distance, IEnumerable<EpicLootRarity> rarities) => builder.SetCondition(new ConditionNearbyPlayerCarryItemWithRarity(distance, rarities));
+    private static void SetConditionItemRarity(ILocalSpawnBuilder builder, int distance, params EpicLootRarity[] rarities) => builder.SetCondition(new ConditionNearbyPlayerCarryItemWithRarity(distance, rarities));
+
+
     public static ILocalSpawnBuilder SetEpicLootConditionNearbyPlayerCarryLegendaryItem(this ILocalSpawnBuilder builder, int distanceToSearch, IEnumerable<string> legendaryItemIds)
     {
         if (IntegrationManager.InstalledEpicLoot)
         {
-            builder.SetCondition(new ConditionNearbyPlayerCarryLegendaryItem(distanceToSearch, legendaryItemIds));
+            SetConditionLegendaryItem(builder, distanceToSearch, legendaryItemIds);
         }
 
         return builder;
@@ -61,9 +67,12 @@ public static class ILocalSpawnBuilderELConditionsExtensions
     {
         if (IntegrationManager.InstalledEpicLoot)
         {
-            builder.SetCondition(new ConditionNearbyPlayerCarryLegendaryItem(distanceToSearch, legendaryItemIds));
+            SetConditionLegendaryItem(builder, distanceToSearch, legendaryItemIds);
         }
 
         return builder;
     }
+
+    private static void SetConditionLegendaryItem(ILocalSpawnBuilder builder, int distanceToSearch, IEnumerable<string> legendaryItemIds) => builder.SetCondition(new ConditionNearbyPlayerCarryLegendaryItem(distanceToSearch, legendaryItemIds));
+    private static void SetConditionLegendaryItem(ILocalSpawnBuilder builder, int distanceToSearch, params string[] legendaryItemIds) => builder.SetCondition(new ConditionNearbyPlayerCarryLegendaryItem(distanceToSearch, legendaryItemIds));
 }
