+++ src/SpawnThat/Integrations/EpicLoot/Models/EpicLootRarity.cs
@@ -1,6 +1,4 @@
-﻿using EpicLoot;
-
-namespace SpawnThat.Integrations.EpicLoot.Models;
+﻿namespace SpawnThat.Integrations.EpicLoot.Models;
 
 public enum EpicLootRarity
 {
@@ -8,25 +6,4 @@ public enum EpicLootRarity
     Rare,
     Epic,
     Legendary,
-}
-
-internal static class EpicLootRarityExtensions
-{
-    public static ItemRarity ToRarity(this EpicLootRarity rarity)
-        => rarity switch
-        {
-            EpicLootRarity.Magic => ItemRarity.Magic,
-            EpicLootRarity.Rare => ItemRarity.Rare,
-            EpicLootRarity.Epic => ItemRarity.Epic,
-            EpicLootRarity.Legendary => ItemRarity.Legendary,
-        };
-
-    public static EpicLootRarity ToRarity(this ItemRarity rarity)
-        => rarity switch
-        {
-            ItemRarity.Magic => EpicLootRarity.Magic,
-            ItemRarity.Rare => EpicLootRarity.Rare,
-            ItemRarity.Epic => EpicLootRarity.Epic,
-            ItemRarity.Legendary => EpicLootRarity.Legendary,
-        };
 }
