+++ src/SpawnThat/Integrations/MobAi/Extensions/IWorldSpawnBuilderMobAiExtensions.cs
@@ -9,9 +9,11 @@ public static class IWorldSpawnBuilderMobAiExtensions
     {
         if (IntegrationManager.InstalledMobAI)
         {
-            builder.SetModifier(new ModifierSetAI(aiName, mobAiConfig));
+            SetModifier(builder, aiName, mobAiConfig);
         }
 
         return builder;
     }
+
+    private static void SetModifier(IWorldSpawnBuilder builder, string aiName, string mobAiConfig) => builder.SetModifier(new ModifierSetAI(aiName, mobAiConfig));
 }
